% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_kkmeans.R
\name{cluster_new}
\alias{cluster_new}
\title{Classify new data based on kkmeans result}
\usage{
cluster_new(test_data, train_data, train_result)
}
\arguments{
\item{test_data}{the new data to be classified}

\item{train_data}{the data to make classifications from}

\item{train_result}{\code{kkmeans} result containing the result from classifying
\code{train_data}}
}
\value{
A vector of class labels for \code{test_data} corresponding to the
clusters present in \code{train_result}.
}
\description{
Given training data, test data, and kkmeans-results,
get which partitions a new set of data belong to
}
