% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mknn_dist.R
\name{get_mknn_dist}
\alias{get_mknn_dist}
\title{Get the average distance to each points k-nearest neighbor}
\usage{
get_mknn_dist(data, k = F)
}
\arguments{
\item{data}{the data vector}

\item{k}{which neighbor to average over}
}
\description{
Given a dataset and a value k, will return the value of the
average distance from each point to it's k-nearest neighbor
}
