% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_kkmeans.R
\name{cluster_new}
\alias{cluster_new}
\title{Classify new data based on kkmeans result}
\usage{
cluster_new(
  test_data,
  train_data,
  train_labels,
  kern = "g",
  param = 1,
  param2 = 1
)
}
\arguments{
\item{kern}{the kernel to use, one of ('gaussian', 'poly'), can use first
letter}

\item{param}{parameter to pass to kernel function.}

\item{data}{data vector}
}
\description{
Given training data, test data, and kkmeans-results,
get which partitions a new set of data belong to
}
