% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kkmeans}
\alias{kkmeans}
\title{An Efficient Kernel K-Means Algorithm}
\usage{
kkmeans(data, k, kern, param, iter_max = 1000L)
}
\arguments{
\item{data}{data to cluster}

\item{k}{the number of clusters}

\item{kern}{the kernel to use, one of ('gaussian', 'poly'), can use first
letter}

\item{params}{parameters to pass to kernel function.}
}
\description{
Performs kernel k-means with the specified kernel using an
algorithm similar to Hartigan and Wong's k-means algorithm.
}
